requirements.txt
discord-py-slash-command==1.0.8.5
discord.py==1.6.0
python-dotenv==0.15.0
